requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI integration
openai>=1.30.0

# Tunnel management
pyngrok==7.0.5

# System monitoring
psutil==5.9.8

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Google Cloud (optional)
google-cloud-storage==2.14.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
